Split CHECK internals to their own files for clean namespacing.
